Move files, fix folder name
