+++ packages/eslint-plugin-query/src/__tests__/stable-query-client.test.ts
@@ -1,6 +1,6 @@
 import { RuleTester } from '@typescript-eslint/rule-tester'
-import { normalizeIndent } from '../../utils/test-utils'
-import { rule } from './stable-query-client.rule'
+import { rule } from '../rules/stable-query-client/stable-query-client.rule'
+import { normalizeIndent } from './test-utils'
 
 const ruleTester = new RuleTester({
   parser: '@typescript-eslint/parser',
