+++ packages/eslint-plugin-query/src/rules.ts
@@ -1,6 +1,6 @@
-import * as exhaustiveDeps from './rules/exhaustive-deps.rule'
+import * as exhaustiveDeps from './rules/exhaustive-deps/exhaustive-deps.rule'
 import * as stableQueryClient from './rules/stable-query-client/stable-query-client.rule'
-import * as noRestDestructuring from './rules/no-rest-desctructuring/no-rest-destructuring.rule'
+import * as noRestDestructuring from './rules/no-rest-destructuring/no-rest-destructuring.rule'
 import type { ESLintUtils } from '@typescript-eslint/utils'
 
 export const rules: Record<
