+++ packages/eslint-plugin-query/src/rules/exhaustive-deps/exhaustive-deps.utils.ts
@@ -1,5 +1,5 @@
 import { AST_NODE_TYPES } from '@typescript-eslint/utils'
-import { ASTUtils } from '../utils/ast-utils'
+import { ASTUtils } from '../../utils/ast-utils'
 import type { TSESLint } from '@typescript-eslint/utils'
 
 export const ExhaustiveDepsUtils = {
