+++ packages/eslint-plugin-query/src/rules/no-rest-destructuring/no-rest-destructuring.rule.ts
